Refactor layout header into modular navigation components
