+++ client/src/components/Layout.tsx
@@ -1,76 +1,19 @@
 import { Link, useLocation } from "wouter";
-import { useState, useEffect, useRef, useCallback, lazy, Suspense } from "react";
-import { Menu, X, Phone, Heart, BookOpen, MessageCircle, Shield, Sparkles, Download, Search as SearchIcon, TrendingUp, Users, ChevronDown, FileText, HelpCircle, BookMarked, Building2, Stethoscope, FolderOpen } from "lucide-react";
+import { useState, useEffect, lazy, Suspense } from "react";
+import { Phone } from "lucide-react";
 const Search = lazy(() => import("@/components/Search"));
 import { Button } from "@/components/ui/button";
-import { motion, AnimatePresence } from "framer-motion";
 import { ScrollToTopButton, Breadcrumbs } from "@/components/UXEnhancements";
+import { HeaderNav } from "@/components/layout/HeaderNav";
+import { navItems, ressourcenItems } from "@/components/layout/navigationData";
 
 interface LayoutProps {
   children: React.ReactNode;
 }
 
-const navItems = [
-  { href: "/verstehen", label: "Verstehen", icon: BookOpen },
-  { href: "/unterstuetzen/uebersicht", label: "Unterstützen", icon: Heart },
-  { href: "/kommunizieren", label: "Kommunizieren", icon: MessageCircle },
-  { href: "/grenzen", label: "Grenzen", icon: Shield },
-  { href: "/selbstfuersorge", label: "Selbstfürsorge", icon: Sparkles },
-  { href: "/genesung", label: "Genesung", icon: TrendingUp },
-  { href: "/beratung", label: "Beratung", icon: Users },
-];
-
-const ressourcenItems = [
-  { href: "/soforthilfe", label: "Soforthilfe", icon: Phone },
-  { href: "/materialien", label: "Materialien & Handouts", icon: Download },
-  { href: "/beratung", label: "Beratung & Netzwerke", icon: Users },
-  { href: "/fachstelle", label: "Fachstelle & Kontakt", icon: Building2 },
-  { href: "/faq", label: "Häufige Fragen (FAQ)", icon: HelpCircle },
-  { href: "/glossar", label: "Glossar", icon: BookMarked },
-  { href: "/buchempfehlungen", label: "Buchempfehlungen", icon: BookOpen },
-  { href: "/unterstuetzen/therapie#therapieangebote", label: "Therapieangebote Zürich", icon: Stethoscope },
-];
-
 export default function Layout({ children }: LayoutProps) {
   const [location] = useLocation();
-  const [mobileMenuOpen, setMobileMenuOpen] = useState(false);
   const [searchOpen, setSearchOpen] = useState(false);
-  const [ressourcenOpen, setRessourcenOpen] = useState(false);
-  const [mobileRessourcenOpen, setMobileRessourcenOpen] = useState(false);
-  const ressourcenRef = useRef<HTMLDivElement>(null);
-  const ressourcenButtonRef = useRef<HTMLButtonElement>(null);
-  const ressourcenTimeoutRef = useRef<ReturnType<typeof setTimeout> | null>(null);
-  const ressourcenOpenRef = useRef(false);
-  const menuItemsRef = useRef<(HTMLAnchorElement | null)[]>([]);
-  const focusedIndexRef = useRef(-1);
-
-  // Sync ref with state so the keydown handler always sees current value
-  useEffect(() => {
-    ressourcenOpenRef.current = ressourcenOpen;
-    if (!ressourcenOpen) {
-      focusedIndexRef.current = -1;
-    }
-  }, [ressourcenOpen]);
-
-  // Focus a menu item by index (clamped to valid range)
-  const focusMenuItem = useCallback((index: number) => {
-    const items = menuItemsRef.current.filter(Boolean) as HTMLAnchorElement[];
-    if (items.length === 0) return;
-    const clamped = ((index % items.length) + items.length) % items.length;
-    focusedIndexRef.current = clamped;
-    items[clamped]?.focus();
-  }, []);
-
-  // Open dropdown and focus first item (WAI-ARIA pattern)
-  const openAndFocusFirst = useCallback(() => {
-    setRessourcenOpen(true);
-    // requestAnimationFrame ensures the menu DOM is rendered before focusing
-    requestAnimationFrame(() => {
-      requestAnimationFrame(() => {
-        focusMenuItem(0);
-      });
-    });
-  }, [focusMenuItem]);
 
   // Keyboard shortcut for search (Ctrl/Cmd + K) + ESC closes dropdown
   useEffect(() => {
@@ -79,98 +22,10 @@ export default function Layout({ children }: LayoutProps) {
         e.preventDefault();
         setSearchOpen(true);
       }
-      if (e.key === "Escape" && ressourcenOpenRef.current) {
-        e.stopPropagation();
-        setRessourcenOpen(false);
-        // Fokus zurück auf den Trigger-Button (A11y-Standard)
-        ressourcenButtonRef.current?.focus();
-      }
     };
     document.addEventListener("keydown", handleKeyDown);
     return () => document.removeEventListener("keydown", handleKeyDown);
   }, []);
-
-  // Keyboard navigation within the open dropdown (Arrow keys, Home, End, Focus-Trap)
-  const handleMenuKeyDown = useCallback((e: React.KeyboardEvent) => {
-    const items = menuItemsRef.current.filter(Boolean) as HTMLAnchorElement[];
-    if (items.length === 0) return;
-
-    switch (e.key) {
-      case "ArrowDown":
-        e.preventDefault();
-        focusMenuItem(focusedIndexRef.current + 1);
-        break;
-      case "ArrowUp":
-        e.preventDefault();
-        focusMenuItem(focusedIndexRef.current - 1);
-        break;
-      case "Home":
-        e.preventDefault();
-        focusMenuItem(0);
-        break;
-      case "End":
-        e.preventDefault();
-        focusMenuItem(items.length - 1);
-        break;
-      case "Tab": {
-        // Focus-Trap: wrap around instead of leaving the menu
-        e.preventDefault();
-        if (e.shiftKey) {
-          // Shift+Tab on first item → last item
-          focusMenuItem(focusedIndexRef.current - 1);
-        } else {
-          // Tab on last item → first item
-          focusMenuItem(focusedIndexRef.current + 1);
-        }
-        break;
-      }
-      default:
-        break;
-    }
-  }, [focusMenuItem]);
-
-  // Close dropdown on outside click
-  useEffect(() => {
-    const handleClickOutside = (e: MouseEvent) => {
-      if (ressourcenRef.current && !ressourcenRef.current.contains(e.target as Node)) {
-        setRessourcenOpen(false);
-      }
-    };
-    document.addEventListener("mousedown", handleClickOutside);
-    return () => document.removeEventListener("mousedown", handleClickOutside);
-  }, []);
-
-  // Close dropdown on route change
-  useEffect(() => {
-    setRessourcenOpen(false);
-    setMobileRessourcenOpen(false);
-  }, [location]);
-
-  // Set data-attribute on body when mobile menu is open (for CSS-based FAB hiding)
-  useEffect(() => {
-    if (mobileMenuOpen) {
-      document.body.setAttribute('data-mobile-menu', 'open');
-    } else {
-      document.body.removeAttribute('data-mobile-menu');
-    }
-    return () => document.body.removeAttribute('data-mobile-menu');
-  }, [mobileMenuOpen]);
-
-  const handleRessourcenEnter = () => {
-    if (ressourcenTimeoutRef.current) {
-      clearTimeout(ressourcenTimeoutRef.current);
-      ressourcenTimeoutRef.current = null;
-    }
-    setRessourcenOpen(true);
-  };
-
-  const handleRessourcenLeave = () => {
-    ressourcenTimeoutRef.current = setTimeout(() => {
-      setRessourcenOpen(false);
-    }, 200);
-  };
-
-  const isRessourcenActive = ressourcenItems.some(item => location.startsWith(item.href.split('#')[0]));
 
   return (
     <div className="min-h-screen flex flex-col bg-background">
@@ -182,300 +37,7 @@ export default function Layout({ children }: LayoutProps) {
         Zum Inhalt springen
       </a>
 
-      {/* Header */}
-      <header className="sticky top-0 z-50 bg-background/95 backdrop-blur-sm border-b border-border/50">
-        <div className="w-full max-w-screen-2xl mx-auto px-4 sm:px-6 lg:px-10">
-          <div className="flex items-center justify-between gap-2 lg:gap-4 h-16 md:h-20">
-            {/* Logo */}
-            <Link href="/" className="flex items-center gap-2 group shrink-0">
-              <img
-                src="https://files.manuscdn.com/user_upload_by_module/session_file/310419663031008193/NqViLFGTREhdSTsm.webp"
-                alt="Startseite"
-                className="w-8 h-8 md:w-10 md:h-10 rounded-full"
-                width={40}
-                height={40}
-                loading="eager"
-                decoding="async"
-              />
-              <span className="hidden 2xl:inline font-medium text-sm text-muted-foreground group-hover:text-foreground transition-colors whitespace-nowrap">
-                Borderline · Hilfe für Angehörige
-              </span>
-            </Link>
-
-            {/* Desktop Navigation */}
-            <nav className="hidden lg:flex items-center gap-0.5 xl:gap-1 shrink-0">
-              {navItems.map((item) => {
-                const isActive = location.startsWith(item.href);
-                return (
-                  <Link
-                    key={item.href}
-                    href={item.href}
-                    className={`px-2.5 lg:px-3 xl:px-3.5 py-2 rounded-lg text-sm font-medium transition-all duration-500 whitespace-nowrap ${
-                      isActive
-                        ? "bg-terracotta-light text-terracotta-darker"
-                        : "text-muted-foreground hover:text-foreground hover:bg-muted"
-                    }`}
-                  >
-                    {item.label}
-                  </Link>
-                );
-              })}
-
-              {/* Ressourcen Dropdown */}
-              <div 
-                ref={ressourcenRef}
-                className="relative"
-                onMouseEnter={handleRessourcenEnter}
-                onMouseLeave={handleRessourcenLeave}
-              >
-                <button
-                  type="button"
-                  ref={ressourcenButtonRef}
-                  onClick={() => ressourcenOpen ? setRessourcenOpen(false) : openAndFocusFirst()}
-                  onKeyDown={(e) => {
-                    if (e.key === "ArrowDown") {
-                      e.preventDefault();
-                      openAndFocusFirst();
-                    }
-                  }}
-                  className={`flex items-center gap-1 px-2.5 lg:px-3 xl:px-3.5 py-2 rounded-lg text-sm font-medium transition-all duration-500 whitespace-nowrap ${
-                    isRessourcenActive || ressourcenOpen
-                      ? "bg-terracotta-light text-terracotta-darker"
-                      : "text-muted-foreground hover:text-foreground hover:bg-muted"
-                  }`}
-                  aria-expanded={ressourcenOpen}
-                  aria-haspopup="true"
-                  aria-controls="ressourcen-menu"
-                >
-                  Ressourcen
-                  <ChevronDown className={`w-3.5 h-3.5 transition-transform duration-300 ${ressourcenOpen ? "rotate-180" : ""}`} />
-                </button>
-
-                <AnimatePresence>
-                  {ressourcenOpen && (
-                    <motion.div
-                      initial={{ opacity: 0, y: -4, scale: 0.98 }}
-                      animate={{ opacity: 1, y: 0, scale: 1 }}
-                      exit={{ opacity: 0, y: -4, scale: 0.98 }}
-                      transition={{ duration: 0.15, ease: "easeOut" }}
-                      id="ressourcen-menu"
-                      role="menu"
-                      aria-label="Ressourcen-Navigation"
-                      onKeyDown={handleMenuKeyDown}
-                      className="absolute right-0 top-full mt-1 w-64 bg-background border border-border/60 rounded-xl shadow-lg shadow-black/8 overflow-hidden z-50"
-                    >
-                      <div className="py-2">
-                        {ressourcenItems.map((item, index) => {
-                          const Icon = item.icon;
-                          const isActive = location === item.href.split('#')[0] || location.startsWith(item.href.split('#')[0] + '/');
-                          const isSoforthilfe = item.href === "/soforthilfe";
-                          return (
-                            <div key={item.href}>
-                              {/* Trennlinie nach Soforthilfe */}
-                              {index === 1 && (
-                                <div className="mx-3 my-1 border-t border-border/40" />
-                              )}
-                              <Link
-                                href={item.href}
-                                role="menuitem"
-                                tabIndex={-1}
-                                ref={(el: HTMLAnchorElement | null) => { menuItemsRef.current[index] = el; }}
-                                onFocus={() => { focusedIndexRef.current = index; }}
-                                onClick={() => setRessourcenOpen(false)}
-                                className={`flex items-center gap-3 px-4 py-2.5 text-sm transition-all outline-none focus-visible:ring-2 focus-visible:ring-terracotta/40 focus-visible:ring-inset ${
-                                  isSoforthilfe
-                                    ? "text-alert font-medium hover:bg-alert/8 focus:bg-alert/8"
-                                    : isActive
-                                      ? "bg-terracotta-light/50 text-terracotta-darker font-medium"
-                                      : "text-muted-foreground hover:text-foreground hover:bg-muted/60 focus:text-foreground focus:bg-muted/60"
-                                }`}
-                              >
-                                <Icon className={`w-4 h-4 shrink-0 ${isSoforthilfe ? "text-alert" : ""}`} />
-                                {item.label}
-                              </Link>
-                            </div>
-                          );
-                        })}
-                      </div>
-                    </motion.div>
-                  )}
-                </AnimatePresence>
-              </div>
-            </nav>
-
-            {/* Search, Emergency Button & Mobile Menu */}
-            <div className="flex items-center gap-2 shrink-0">
-              {/* Search Button – Desktop */}
-              <button
-                type="button"
-                onClick={() => setSearchOpen(true)}
-                className="hidden sm:flex items-center gap-2 px-3 py-1.5 rounded-lg border border-border/50 text-muted-foreground hover:text-foreground hover:border-border transition-colors text-sm"
-                aria-label="Suchen"
-              >
-                <SearchIcon className="w-4 h-4" />
-                <span>Suchen</span>
-              </button>
-              {/* Search Button – Mobile (nur Icon) */}
-              <button
-                type="button"
-                onClick={() => setSearchOpen(true)}
-                className="sm:hidden p-2.5 -m-0.5 rounded-lg text-muted-foreground hover:text-foreground hover:bg-muted transition-all"
-                aria-label="Suche öffnen"
-              >
-                <SearchIcon className="w-5 h-5" />
-              </button>
-
-              <Link href="/soforthilfe" aria-label="Soforthilfe – Notfallnummern und Krisenberatung">
-                <Button
-                  variant="default"
-                  size="sm"
-                  className="bg-alert hover:bg-alert/85 text-white hidden sm:flex"
-                  tabIndex={-1}
-                >
-                  <Phone className="w-4 h-4 lg:mr-0 xl:mr-2" />
-                  <span className="hidden xl:inline">Soforthilfe</span>
-                </Button>
-              </Link>
-              
-              {/* Mobile Menu Button */}
-              <button
-                type="button"
-                onClick={() => setMobileMenuOpen(!mobileMenuOpen)}
-                className="lg:hidden p-2.5 -m-0.5 rounded-lg hover:bg-muted transition-colors"
-                aria-label={mobileMenuOpen ? "Menü schliessen" : "Menü öffnen"}
-              >
-                {mobileMenuOpen ? <X className="w-6 h-6" /> : <Menu className="w-6 h-6" />}
-              </button>
-            </div>
-          </div>
-        </div>
-
-        {/* Mobile Navigation */}
-        <AnimatePresence>
-          {mobileMenuOpen && (
-            <motion.div
-              initial={{ opacity: 0, height: 0 }}
-              animate={{ opacity: 1, height: "auto" }}
-              exit={{ opacity: 0, height: 0 }}
-              transition={{ duration: 0.5, ease: "easeOut" }}
-              className="lg:hidden border-t border-border/50 bg-background overflow-y-auto overscroll-contain"
-              style={{
-                maxHeight: 'calc(100dvh - 5rem)',
-                WebkitOverflowScrolling: 'touch',
-              }}
-              onKeyDown={(e) => {
-                if (e.key === "Escape") setMobileMenuOpen(false);
-              }}
-            >
-              <nav className="container py-4 flex flex-col gap-2" style={{ paddingBottom: 'calc(16px + env(safe-area-inset-bottom, 0px) + 88px)' }}>
-                {navItems.map((item) => {
-                  const Icon = item.icon;
-                  const isActive = location.startsWith(item.href);
-                  return (
-                    <Link
-                      key={item.href}
-                      href={item.href}
-                      onClick={() => setMobileMenuOpen(false)}
-                      className={`flex items-center gap-3 px-4 py-3.5 rounded-lg text-base font-medium transition-all ${
-                        isActive
-                          ? "bg-terracotta-light text-terracotta-darker"
-                          : "text-muted-foreground hover:text-foreground hover:bg-muted"
-                      }`}
-                    >
-                      <Icon className="w-5 h-5" />
-                      {item.label}
-                    </Link>
-                  );
-                })}
-
-                {/* Mobile Ressourcen Accordion */}
-                <div className="mt-1">
-                  <button
-                    type="button"
-                    onClick={() => setMobileRessourcenOpen(!mobileRessourcenOpen)}
-                    aria-expanded={mobileRessourcenOpen}
-                    aria-controls="mobile-ressourcen-menu"
-                    className={`flex items-center justify-between w-full px-4 py-3 rounded-lg text-base font-medium transition-all ${
-                      isRessourcenActive
-                        ? "bg-terracotta-light text-terracotta-darker"
-                        : "text-muted-foreground hover:text-foreground hover:bg-muted"
-                    }`}
-                  >
-                    <span className="flex items-center gap-3">
-                      <FolderOpen className="w-5 h-5" />
-                      Ressourcen
-                    </span>
-                    <ChevronDown className={`w-4 h-4 transition-transform duration-300 ${mobileRessourcenOpen ? "rotate-180" : ""}`} />
-                  </button>
-                  
-                  <AnimatePresence>
-                    {mobileRessourcenOpen && (
-                      <motion.div
-                        initial={{ opacity: 0, height: 0 }}
-                        animate={{ opacity: 1, height: "auto" }}
-                        exit={{ opacity: 0, height: 0 }}
-                        transition={{ duration: 0.3, ease: "easeOut" }}
-                        className="overflow-hidden"
-                      >
-                        <div id="mobile-ressourcen-menu" role="menu" aria-label="Ressourcen-Navigation" className="pl-4 mt-1 flex flex-col gap-1">
-                          {ressourcenItems.map((item) => {
-                            const Icon = item.icon;
-                            const isActive = location === item.href.split('#')[0];
-                            const isSoforthilfe = item.href === "/soforthilfe";
-                            return (
-                              <Link
-                                key={item.href}
-                                href={item.href}
-                                role="menuitem"
-                                onClick={() => {
-                                  setMobileMenuOpen(false);
-                                  setMobileRessourcenOpen(false);
-                                }}
-                                className={`flex items-center gap-3 px-4 py-2.5 rounded-lg text-sm font-medium transition-all ${
-                                  isSoforthilfe
-                                    ? "text-alert font-semibold"
-                                    : isActive
-                                      ? "bg-terracotta-light/50 text-terracotta-darker"
-                                      : "text-muted-foreground hover:text-foreground hover:bg-muted"
-                                }`}
-                              >
-                                <Icon className={`w-4 h-4 ${isSoforthilfe ? "text-alert" : ""}`} />
-                                {item.label}
-                              </Link>
-                            );
-                          })}
-                        </div>
-                      </motion.div>
-                    )}
-                  </AnimatePresence>
-                </div>
-
-                <Link
-                  href="/soforthilfe"
-                  onClick={() => setMobileMenuOpen(false)}
-                  className="flex items-center gap-3 px-4 py-3 rounded-lg text-base font-medium bg-alert text-white mt-2"
-                >
-                  <Phone className="w-5 h-5" />
-                  Soforthilfe
-                </Link>
-                {/* Mobile Search Button */}
-                <button
-                  type="button"
-                  onClick={() => {
-                    setMobileMenuOpen(false);
-                    setSearchOpen(true);
-                  }}
-                  className="flex items-center gap-3 px-4 py-3 rounded-lg text-base font-medium text-muted-foreground hover:text-foreground hover:bg-muted mt-2 border border-border/50"
-                  aria-label="Suche öffnen"
-                >
-                  <SearchIcon className="w-5 h-5" />
-                  Suchen
-                </button>
-              </nav>
-            </motion.div>
-          )}
-        </AnimatePresence>
-      </header>
+      <HeaderNav onSearchOpen={() => setSearchOpen(true)} />
 
       <div className="border-b border-border/40 bg-sage-wash/40">
         <div className="w-full max-w-screen-2xl mx-auto px-4 sm:px-6 lg:px-10 py-1.5 text-xs text-muted-foreground flex flex-wrap items-center gap-x-3 gap-y-1">
