+++ client/src/components/layout/navigationData.ts
@@ -1,0 +1,29 @@
+import { BookMarked, BookOpen, Building2, Download, Heart, HelpCircle, MessageCircle, Phone, Shield, Sparkles, Stethoscope, TrendingUp, Users } from "lucide-react";
+import type { LucideIcon } from "lucide-react";
+
+export interface NavItem {
+  href: string;
+  label: string;
+  icon: LucideIcon;
+}
+
+export const navItems: NavItem[] = [
+  { href: "/verstehen", label: "Verstehen", icon: BookOpen },
+  { href: "/unterstuetzen/uebersicht", label: "Unterstützen", icon: Heart },
+  { href: "/kommunizieren", label: "Kommunizieren", icon: MessageCircle },
+  { href: "/grenzen", label: "Grenzen", icon: Shield },
+  { href: "/selbstfuersorge", label: "Selbstfürsorge", icon: Sparkles },
+  { href: "/genesung", label: "Genesung", icon: TrendingUp },
+  { href: "/beratung", label: "Beratung", icon: Users },
+];
+
+export const ressourcenItems: NavItem[] = [
+  { href: "/soforthilfe", label: "Soforthilfe", icon: Phone },
+  { href: "/materialien", label: "Materialien & Handouts", icon: Download },
+  { href: "/beratung", label: "Beratung & Netzwerke", icon: Users },
+  { href: "/fachstelle", label: "Fachstelle & Kontakt", icon: Building2 },
+  { href: "/faq", label: "Häufige Fragen (FAQ)", icon: HelpCircle },
+  { href: "/glossar", label: "Glossar", icon: BookMarked },
+  { href: "/buchempfehlungen", label: "Buchempfehlungen", icon: BookOpen },
+  { href: "/unterstuetzen/therapie#therapieangebote", label: "Therapieangebote Zürich", icon: Stethoscope },
+];
